refactor: main 변경 사항 반영
