+++ src/view/index/Card/index.css.ts
@@ -9,6 +9,7 @@ export const cardStyle = style({
   gap: "2rem",
 
   width: "40rem",
+  height: "31rem",
   padding: "0 3rem 3rem",
 
   borderRadius: "2.4rem",
